Search for matches to sequences against the BOLD COI database.

Search for matches to sequences against the BOLD COI database.

bold_identify(sequences, db = "COX1", response = FALSE, ...)

Arguments

sequences
(character) Returns all records containing matching marker codes. Required. One or more. See Details.

db
(character) The database to match against, one of COX1, COX1_SPECIES, COX1_SPECIES_PUBLIC, OR COX1_L604bp. See Details for more information.

response
(logical) Note that response is the object that returns from the Curl call, useful for debugging, and getting detailed info on the API call.

...
Further args passed on to crul::verb-GET, main purpose being curl debugging

BOLD only allows one sequence per query. We internally lapply over the input values given to the sequences` parameter to search with one sequence per query. Remember this if you have a lot of sequences - you are doing a separate query for each one, so it can take a long time - if you run into errors let us know.

Value

A data.frame with details for each specimen matched. if a failed request, returns NULL

db parmeter options

COX1 Every COI barcode record on BOLD with a minimum sequence length of 500bp (warning: unvalidated library and includes records without species level identification). This includes many species represented by only one or two specimens as well as all species with interim taxonomy. This search only returns a list of the nearest matches and does not provide a probability of placement to a taxon.

COX1_SPECIES Every COI barcode record with a species level identification and a minimum sequence length of 500bp. This includes many species represented by only one or two specimens as well as all species with interim taxonomy.

COX1_SPECIES_PUBLIC All published COI records from BOLD and GenBank with a minimum sequence length of 500bp. This library is a collection of records from the published projects section of BOLD.

OR COX1_L604bp Subset of the Species library with a minimum sequence length of 640bp and containing both public and private records. This library is intended for short sequence identification as it provides maximum overlap with short reads from the barcode region of COI.

Named outputs

To maintain names on the output list of data make sure to pass in a named list to the sequences parameter. You can for example, take a list of sequences, and use stats::setNames() to set names.

References

http://v4.boldsystems.org/index.php/resources/api?type=idengine

See also

Examples

if (FALSE) {
seq <- sequences$seq1
res <- bold_identify(sequences=seq)
head(res[[1]])
head(bold_identify(sequences=seq, db='COX1_SPECIES')[[1]])
}